Order write updates

+++ OrderService/src/main/java/com/publicissapient/DAO/OrderDAO.java
@@ -3,7 +3,7 @@ package com.publicissapient.DAO;
 import com.publicissapient.handler.ResourceNotFoundException;
 
 public interface OrderDAO {
-	public String getOrderByOrderId(Integer orderId) throws ResourceNotFoundException;
+	public String getOrderByOrderId(long orderId) throws ResourceNotFoundException;
 	
 
 }
